# lib/api-zod/src/generated/types/analysis.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * InsightForge AI data intelligence API
 * OpenAPI spec version: 0.1.0
 */
import type { AnalysisPlan } from './analysisPlan';
import type { AnalysisStep } from './analysisStep';
import type { AnalysisSummary } from './analysisSummary';
import type { Chart } from './chart';
import type { Evidence } from './evidence';
import type { Insight } from './insight';
import type { QualityDetails } from './qualityDetails';
import type { Simulation } from './simulation';

export type Analysis = AnalysisSummary & {
  plan: AnalysisPlan;
  steps: AnalysisStep[];
  insights: Insight[];
  charts: Chart[];
  qualityDetails: QualityDetails;
  evidence: Evidence[];
  geographic: boolean;
  whatIf: Simulation;
};